+++ admin-dev/themes/new-theme/js/pages/product/mixins/is-attribute-selected.ts
@@ -22,7 +22,7 @@
  * @copyright Since 2007 PrestaShop SA and Contributors
  * @license   https://opensource.org/licenses/OSL-3.0 Open Software License (OSL 3.0)
  */
-import {Attribute, AttributeGroup} from '@pages/product/components/generator';
+import {Attribute, AttributeGroup} from '@pages/product/types';
 
 export default {
   methods: {
