+++ admin-dev/themes/new-theme/js/types/renderers/index.d.ts
