+++ admin-dev/themes/new-theme/js/types/showcase/index.d.ts
