+++ admin-dev/themes/new-theme/js/types/vue/shim.d.ts
@@ -1,3 +1,5 @@
+// Makes TS happy because it doesn't know about Vue SFC (not like webpack and the loader)
+// so it will treat *.vue files as defineComponent
 declare module '*.vue' {
   import {defineComponent} from 'vue';
